Storing of the alerts working

+++ BrickedUp/database/migrations/2024_12_15_121039_add_user_id_to_price_alerts_table.php
@@ -12,20 +12,21 @@ return new class extends Migration
     public function up()
     {
         Schema::table('price_alerts', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->after('id');
-    
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            if (!Schema::hasColumn('price_alerts', 'user_id')) {
+                $table->unsignedBigInteger('user_id')->after('id');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down()
     {
         Schema::table('price_alerts', function (Blueprint $table) {
-            $table->dropForeign(['user_id']);
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('price_alerts', 'user_id')) {
+                $table->dropForeign(['user_id']);
+                $table->dropColumn('user_id');
+            }
         });
     }
+
 };
